開始編寫PWM部分

+++ SYSTEM/Misc/timing_delay.c
@@ -71,38 +71,38 @@ static __IO uint32_t TimingDelay;
   */
 void TimingDelay_Init(void)
 {
-    /* Enable TIM2 clock */
-    CLK_PeripheralClockConfig(CLK_Peripheral_TIM2, ENABLE);
+  /* Enable TIM2 clock */
+  CLK_PeripheralClockConfig(CLK_Peripheral_TIM2, ENABLE);
 
-    /* Remap TIM2 ETR to LSE: TIM2 external trigger becomes controlled by LSE clock */
-    SYSCFG_REMAPPinConfig(REMAP_Pin_TIM2TRIGLSE, ENABLE);
+  /* Remap TIM2 ETR to LSE: TIM2 external trigger becomes controlled by LSE clock */
+  SYSCFG_REMAPPinConfig(REMAP_Pin_TIM2TRIGLSE, ENABLE);
 
-    /* Enable LSE clock */
-    CLK_LSEConfig(CLK_LSE_ON);
-    /* Wait for LSERDY flag to be reset */
-    while (CLK_GetFlagStatus(CLK_FLAG_LSERDY) == RESET)
-        ;
+  /* Enable LSE clock */
+  CLK_LSEConfig(CLK_LSE_ON);
+  /* Wait for LSERDY flag to be reset */
+  while (CLK_GetFlagStatus(CLK_FLAG_LSERDY) == RESET)
+    ;
 
-    /* TIM2 configuration:
+  /* TIM2 configuration:
      - TIM2 ETR is mapped to LSE
      - TIM2 counter is clocked by LSE div 4
       so the TIM2 counter clock used is LSE / 4 = 32.768 / 4 = 8.192 KHz
      TIM2 Channel1 output frequency = TIM2CLK / (TIM2 Prescaler * (TIM2_PERIOD + 1))
                                     = 8192 / (1 * 8) = 1024 Hz                */
 
-    /* Time Base configuration */
-    TIM2_TimeBaseInit(TIM2_Prescaler_1, TIM2_CounterMode_Up, TIM2_PERIOD);
-    TIM2_ETRClockMode2Config(TIM2_ExtTRGPSC_DIV4, TIM2_ExtTRGPolarity_NonInverted, 0);
+  /* Time Base configuration */
+  TIM2_TimeBaseInit(TIM2_Prescaler_1, TIM2_CounterMode_Up, TIM2_PERIOD);
+  TIM2_ETRClockMode2Config(TIM2_ExtTRGPSC_DIV4, TIM2_ExtTRGPolarity_NonInverted, 0);
 
-    TIM2_UpdateRequestConfig(TIM2_UpdateSource_Global);
+  TIM2_UpdateRequestConfig(TIM2_UpdateSource_Global);
 
-    /* Clear TIM2 update flag */
-    TIM2_ClearFlag(TIM2_FLAG_Update);
+  /* Clear TIM2 update flag */
+  TIM2_ClearFlag(TIM2_FLAG_Update);
 
-    /* Enable update interrupt */
-    TIM2_ITConfig(TIM2_IT_Update, ENABLE);
+  /* Enable update interrupt */
+  TIM2_ITConfig(TIM2_IT_Update, ENABLE);
 
-    TIM2_Cmd(ENABLE);
+  TIM2_Cmd(ENABLE);
 }
 
 /**
@@ -112,9 +112,9 @@ void TimingDelay_Init(void)
   */
 void Delay(__IO uint32_t nTime)
 {
-    TimingDelay = nTime;
-    while (TimingDelay != 0)
-        ;
+  TimingDelay = nTime;
+  while (TimingDelay != 0)
+    ;
 }
 
 /**
@@ -122,13 +122,13 @@ void Delay(__IO uint32_t nTime)
   * @param  nCount: specifies the delay time length.
   * @retval None
   */
-void Delay_By_nop(__IO uint16_t nCount)
+void Delay_By_nop(__IO uint32_t nCount)
 {
-    /* Decrement nCount value */
-    while (nCount != 0)
-    {
-        nCount--;
-    }
+  /* Decrement nCount value */
+  while (nCount != 0)
+  {
+    nCount--;
+  }
 }
 
 /**
@@ -147,10 +147,10 @@ void Delay_By_nop(__IO uint16_t nCount)
   */
 void TimingDelay_Decrement(void)
 {
-    if (TimingDelay != 0x00)
-    {
-        TimingDelay--;
-    }
+  if (TimingDelay != 0x00)
+  {
+    TimingDelay--;
+  }
 }
 
 /**
